Add support for basic version requirements
With any project, eventually you get to a stage where you have
configuration defined that works on a specific version of your software.
Upgrading between versions usually requires some effort but it's rolled
out when it's needed. This becomes even more difficult working in a team
where people can have different versions of the software that you use
based on how often they like to update their systems.

To handle these scenarios, you introduce some sort of versioning that
defines what versions you want the configuration to operate with and the
users then need to meet those requirements.

This introduces a basic concept of the versioning for IAMy. The logic is
pretty straight forward.

- If you find a `.iamy-version` file, use the version defined in it. If
  not, continue on.
- If the `.iamy-version` file contains a version number, compare it to
  the local version and ensure that the version in use meets the minimum
  requirements to operate. If it doesn't, inform the user of the
  discrepancy and what is required.

#62 proposes quite a few more coditions to restrain the version however
I don't think that's required for majority of use cases. I think
defining a minimum version covers most of the sharp edges we've
encountered. But, we can always make this more perscriptive if needed.

Closes #62